* master: (71 commits)
  Replaced png icons with vectors (#2288)
  Dark theme not applied when user load settings from collect.settings file (#2245)
  Removed unused ic_menu_forward.png icon (#2289)
  Remove toast informing about spaces in underlying values (#2285)
  Fixed SQL injection vulnerability (#2281)
  Collect more database logs
  Fixes issues after rebasing onto master
  Add checkstyle rule to enforce CONSTANT_NAME pattern
  Remove boolean constant from ListMultiWidget Fix constant name
  Add final to make fields immutable
  Update AllWidgetsFormTest to match XLSForm in README (#2256)
  Closes #2269: Added null check for removedTempInstance() (#2277)
  Replace the icons in DrawView with Vectors. (#2282)
  Refactored HierarchyView using RecyclerView (#2128)
  Issue #2264 None translatable string-array time_scale in GeoTrace (#2273)
  Replace the activity we show after clicking on a notification with a dialog (#2236)
  Fixed NullPointerException in settings (#2275)
  Render warning on question widget in red
  Add documentation for SpacesInUnderlyingValuesWarning
  Clean up: remove RelativeLayout leftover qualifier from LinearLayout
  ...

# Conflicts:
#	collect_app/src/main/java/org/odk/collect/android/tasks/InstanceServerUploader.java
#	collect_app/src/main/java/org/odk/collect/android/utilities/ResponseMessageParser.java

+++ collect_app/src/main/java/org/odk/collect/android/widgets/VideoWidget.java
@@ -196,7 +196,7 @@ public class VideoWidget extends QuestionWidget implements FileWidget {
         MediaManager
                 .INSTANCE
                 .markOriginalFileOrDelete(getFormEntryPrompt().getIndex().toString(),
-                getInstanceFolder() + File.separator + binaryName);
+                        getInstanceFolder() + File.separator + binaryName);
         binaryName = null;
     }
 
@@ -218,23 +218,40 @@ public class VideoWidget extends QuestionWidget implements FileWidget {
         }
     }
 
+    /**
+     * Set this widget with the actual file returned by OnActivityResult.
+     * Both of Uri and File are supported.
+     * If the file is local, a Uri is enough for the copy task below.
+     * If the chose file is from cloud(such as Google Drive),
+     * The retrieve and copy task is already executed in the previous step,
+     * so a File object would be presented.
+     *
+     * @param object Uri or File of the chosen file.
+     * @see org.odk.collect.android.activities.FormEntryActivity#onActivityResult(int, int, Intent)
+     */
     @Override
-    public void setBinaryData(Object binaryuri) {
-        if (binaryuri == null || !(binaryuri instanceof Uri)) {
-            Timber.w("AudioWidget's setBinaryData must receive a Uri object.");
+    public void setBinaryData(Object object) {
+        File newVideo = null;
+        // get the file path and create a copy in the instance folder
+        if (object instanceof Uri) {
+            String sourcePath = getSourcePathFromUri((Uri) object);
+            String destinationPath = getDestinationPathFromSourcePath(sourcePath);
+            File source = fileUtil.getFileAtPath(sourcePath);
+            newVideo = fileUtil.getFileAtPath(destinationPath);
+            fileUtil.copyFile(source, newVideo);
+        } else if (object instanceof File) {
+            newVideo = (File) object;
+        } else {
+            Timber.w("VideoWidget's setBinaryData must receive a File or Uri object.");
             return;
         }
 
-        // get the file path and create a copy in the instance folder
-        Uri uri = (Uri) binaryuri;
 
-        String sourcePath = getSourcePathFromUri(uri);
-        String destinationPath = getDestinationPathFromSourcePath(sourcePath);
+        if (newVideo == null) {
+            Timber.e("setBinaryData FAILED");
+            return;
+        }
 
-        File source = fileUtil.getFileAtPath(sourcePath);
-        File newVideo = fileUtil.getFileAtPath(destinationPath);
-
-        fileUtil.copyFile(source, newVideo);
 
         if (newVideo.exists()) {
             // Add the copy to the content provier
@@ -269,10 +286,11 @@ public class VideoWidget extends QuestionWidget implements FileWidget {
         // a bug in the Nexus 7 on 4.3 not returning the mediaUri in the data
         // of the intent - uri in this case is a file
         if (NEXUS7.equals(MODEL) && Build.VERSION.SDK_INT == 18) {
-            File fileToDelete = new File(uri.getPath());
-            int delCount = fileToDelete.delete() ? 1 : 0;
-
-            Timber.i("Deleting original capture of file: %s count: %d", uri.toString(), delCount);
+            if (object instanceof File) {
+                File fileToDelete = (File) object;
+                int delCount = fileToDelete.delete() ? 1 : 0;
+                Timber.i("Deleting original capture of file: %s count: %d", binaryName, delCount);
+            }
         }
     }
